Introduced pluggable data model.
A new extension point ProjectGridBuilder is added so that the data model behind visualization can be created in different ways. The default implementation that uses the downstream/upstream relationship (which is the current behavior) is provided, and the backward compatible data read-in code ensures that existing configuration will work the same way.

The extension point allows other plugins (such as the 'promoted-builds' plugin) to contribute different mechanisms, such as one based on fingerprinting.